Update TServerObserver instance also within the new transports
Summary:
THeader transport layer feeds statistical information to TServerObserver object.
Http2 and RSocket transport layers should also inform the Observer of the ThriftServer, such as if timeout occurs, when a request is received, when a reply is sent etc.

As the Observer of the ThriftServer can be set at any time and the ThriftProcessor and the TransportRoutingHandler needs the most recent Observer, I have added the getObserver() method to the ServerConfigs interface. When instances of these classes need the Observer, they request it from the ServerConfigs.

Currently I have added support for listening Timeout events, I will make all of the events are properly populated.

Reviewed By: yfeldblum, lehecka

Differential Revision: D6543571

fbshipit-source-id: b0b818354f412f2695a8a03516522c732498f203