Container Flat Names need to include program name if external
Summary:
We have a collision with Containers of Structs, where the struct from an
external program and there are multiple structs being used with that base name.

The fix is to add the name of the program to the type in flatten_type_name
when the type is from an external program.

Differential Revision: D6584538

fbshipit-source-id: 42d08f684fe0215176f3658549172cc660201f71