add the abiity to add attributes to thrift generated enums
Summary:
Original reasoning:

the JS Enum attribute allows seamless syncing of enums between hack and thrift. Up until now, the lack of this ability has been a pain point for www devs, since they would have to either manually duplicate the enums or use deprecated methods of enum syncing. This fixes that.

This is just the generic version of that

Reviewed By: yfeldblum

Differential Revision: D6604062

fbshipit-source-id: 3ba4f49c693dd98249b923a88f2dcf9ab7977f3a