Reintroduce D6599248 in a less fatal form
Summary:
Cast port to an int(..) if unix_socket isn't specified for API clarity with strictness (unix_sockets don't have ports), and to ensure that AssertionError's don't get bubbled up and misinterpreted by the caller.

Unfortunately, downstream callers expect `.port` to be populated with a value, so just typecheck in the non-unix_socket case.

This is an alternative to the change proposed in D6611745.

Reviewed By: yfeldblum

Differential Revision: D6614352

fbshipit-source-id: 89450964cb2920b28069779aa66adebd63c481f2