Fix the timeout related unit tests
Summary:
Noticed that when we provide a timeout value that is too low, i.e. 1, to the HHWheelTimer, it never times out at least 1 out of 500 times. So I have increased the timeout value.
It might be related to Bucket'ing logic in the timer. It should not affect the real use cases but it affects the unit tests.

Reviewed By: eduardo-elizondo

Differential Revision: D6610574

fbshipit-source-id: e753a63d06fbf17362f568aeefe3888d6443faed