ClientWrapper to unique_ptr and single copy of the async_client
Summary:
Moved the ClientWrappers to utilize a single ptr in the base class.
Move from shared_ptr to unique_ptr.

added a constructor that can take a channel, so we don't need the helper methods
anymore

This is cleanup for future work to have custom ClientWrapper methods to
better handle return headers and rpcOptions.

Reviewed By: yfeldblum

Differential Revision: D6731282

fbshipit-source-id: 8a78ba1dfb3d1e899070d5e789f3148fad3db540