Perf Enhancement with Enum arguments to Struct constructors and RPC calls
Summary:
Before we were checkin if (python) enum  == (python) enum member.
This is very slow, for one we have to call .value 2 times to find the equivalent
c++ enum_value.  In the worst case the last enum value of a huge enum it was O(2*n)
calls to .value which is not cheap even in python standards because it
uses DynamicClassAttribute decorator in python to do some magic its almost 3x slower
than normal

The replacement is to call .value once on the passed in python enum. Store
that value in a c int, and compare against the value from the .thrift file.

After doing this cython is able to turn the entire if/elif chain into a switch
statement.

made the _to_cpp function inline for good measure.

Reviewed By: ot

Differential Revision: D6864135

fbshipit-source-id: 2af75870800e23c4331c3bc51a5799bb5573b13d