Cross-module typedef example in the compiler test fixtures
Summary:
We want to see how the Thrift compiler generates code from cross-module typedefs.

This is separate from D6947687 to differentiate the "gold output" changes caused by the changes to the compiler code from the "gold output" changes caused by the modifications to the `.thrift` sources made in this diff.

Reviewed By: marcinpe

Differential Revision: D6954709

fbshipit-source-id: d1a0562fbf12feb75d92e8f6de76a210e33356ba