typedefs should provide python type aliases
Summary:
Let's say we have `BigInterface.thrift`, which is used widely, including from Python programs.

If we want to break out part of it into `SubInterface.thrift`, we might add it to `BigInterface.thrift` an include for backwards-compatibility:

```
include "SubInterface.thrift"
```

However, this is not yet backwards-compatible, because any clients of `BigInterface.thrift` would need to be changed to use `SubInterface.SomeType` to access the types we broke out. To be truly transparent, we want something like this in `BigInterface.thrift`:

```
include "SubInterface.thrift"

typedef SubInterface.Thrift.MovedType1 MovedType1
typedef SubInterface.Thrift.MovedType1 MovedType2
```

Before the present change, the Python generator would ignore typedefs, so this kind of refactor would be impossible.

I also added a unit test that verifies that the newly generated Python works as expected.

Reviewed By: yfeldblum

Differential Revision: D6947687

fbshipit-source-id: 604974041eed5c42c68f0d26c3b02febf66de09b