Fix cache key collision for modules with the same name
Summary:
Right now we use program name + namespace as a cache key. This is a bad approach, because you may have files with the same name + namespace, but in different locations and in such cases we get cache key collision and generate wrong output (_types.h would include itself).

This diff changes cache key to file path, so that we don't get any collisions.

Reviewed By: yfeldblum

Differential Revision: D7114925

fbshipit-source-id: d79f99c7739facf11e706aa2aecb0483ad46ff72