Change names of thread pools that server spins up
Summary:
Right now Thrift uses default name for IOThreadPool, which makes it hard to distinguish from thread pools that people create. This changes that to a more distinct name `ThriftIO`.
This also fixes a bug in SSLHandshakePool name - thread name should not exceed 15 characters, and in this case it's 16 chars prefix alone.

Reviewed By: yfeldblum, hercule24

Differential Revision: D7130482

fbshipit-source-id: fb977231454568f09aedd4218a31552500ce64c3