Add new options to the Thrift hack gen that allow for graceful migration from php code gen to hack code gen
Summary:
add no_use_hack_collections and client_accepts_null. The former forces thrift to never use hack collection objects, which it often did for member variables. The latter allows the clients to accept null for arguments and handles it gracefully, as the php code gen does.

At the same time, fix a few minor bugs that were exposed along the way.

For the record, the options utilized for this migration are:
1. array_migration
2. arraysets
3. oldenum
4. no_use_hack_collections
5. mapconstruct
6. client_accepts_null

Reviewed By: yfeldblum

Differential Revision: D7118175

fbshipit-source-id: f756707a4e8630ca022f04e15ebcfaa9f7ae3a95