Iterate collections with auto and cbegin
Summary:
[Thrift] Iterate collections with `auto` and `cbegin`.

There is a fixture which does not compile without using `auto` because the collection type is speciied as `const std::map`, and blindly plugging `const std::map` into the generated code as the type of a local variable gives the wrong AST. Use `auto` to solve that problem. The fixture in question is `thrift/compiler/test/fixtures/refs/src/module.thrift`.

While we are here, use `cbegin` and `cend` to enforce that iteration for serializing a collection does not mutate the collection being iterated, even if the collection from which we get the iterators is not `const`. This is not strictly necessary, but let us do it anyway.

Reviewed By: stevegury

Differential Revision: D3748540

fbshipit-source-id: 40947d2db1b480e956796e745ec4ef3fba4c6bd8