Fixed Long Standing issue with Clients causing aborts when used wrong
Summary:
Python is considered a Safe Enivronment, where things like aborts and segfaults
can not happen.

If clients where used outside of "async with" we use to print a error message
and then C++ would abort.

This fix makes sure C++ no longer aborts but we still print the message
that the "async with" should be used.

Reviewed By: ambv

Differential Revision: D6923956

fbshipit-source-id: 4f59f6c3c8aa5c8b2c279c728a8c011a040940af