Add ConcurrentServer
Summary:
This adds concurrent_server.go which could be used on its own or through thriftbase.ConcurrentServer.

ConcurrentServer is similar to SimpleServer (simple_server.go), which reponds to thrift requests of a generated service type. The difference is that while SimpleServer processes requests on the same connection serially, ConcurrentServer processes can process them in out-of-order fashion to avoid head of line blocking

Reviewed By: fiorix

Differential Revision: D7374124

fbshipit-source-id: 79e9815208bf19d4415d18f50d849ba8b0fa354c