Gracefully handle the already accepted but not registered connections while closing the RSocketServer
Summary:
When the RSocketServer shuts down, it first closes the connection acceptor, which is a DuplexConnectionAcceptor.
It then closes the `setup frame` acceptors, which are the instances of SetupResumeAcceptor.

If a connection is already accepted via DuplexConnectionAcceptor but not yet accepted by SetupResumeAcceptor, then this connection will start waiting for the Setup Frame. In this case, when the `setup frame` arrives the server, the RSocketServer instance will already be deallocated.

 - Updated the RSocketServer code and made the onRSocketSetup function don't depend on the RSocketServer instance, via making it a static function.
 - As the ConnectionSet implementation is improved to be safe for such races, I have updated RSocketServer to keep shared_ptr to this class instead of unique_ptr and I have provided a weak_ptr of this data member into the onRSocketSetup lambda function. So I have deferred the responsibility to that class.

These two updates enabled the onRSocketSetup function to further be able to understand if the server is shutting down and it doesn't access to the already deallocated objects anymore.

--

Furthermore, RSRoutingHandler's two methods, handleConnection and stopListening, is called by two different threads.
 - This causes race condition between destruction of RSocketServer and calling acceptConnection method on it.
 - As we have made RSocketServer safe on calling shutdownAndWait and acceptConnection, instead of keeping
 - RSocketServer as unique_ptr, we keep it as shared_ptr at the RSRoutingHandler class.
 - So we also defer the graceful handling of races at the destruction to RSocketServer which defers it to ConnectionSet.

Reviewed By: alexmalyshev

Differential Revision: D7579762

fbshipit-source-id: 3e7fe23e0174c8e20fc79b86bf7fd4dbaab6067a