respect optional fields in debug_equal and pretty_print
Summary:
Previously two structures which were equal except that one side's `__isset` for an `optional` field had a different value would compare equal in `debug_equal`. Now they will be correctly reported as different and the difference will be passed to the output formatter.

Similarly, all fields were printed by `pretty_print`, whereas now fields are not printed if they are `optional` and not set.

Implementation note - I chose to check `member::optional::value` is `optional` as well as `!member::is_set(what)` because `member::is_set(what)` commonly evaluates to false for fields which are not either `optional` or `required` in the IDL. I wanted this code to behave consistently with the serializers which write these fields out whether or not they pass the `is_set` filter.

Reviewed By: yfeldblum, stevegury

Differential Revision: D7282581

fbshipit-source-id: 1f5077196461ed54e03b06ba41b31f224057ae3c