make thrift mstch_pyi namespace match the target namespace
Summary:
In thrift mstch pyi type stub generation, if the last element of the thrift py namespace doesn't match the thrift filename, it appends the thrift filename to the namespace. This is a bug because the type stubs' package should match the py namespace perfectly, otherwise the type stubs are useless.

To fix this, remove the code that appends the filename to the thrift namespace in this case

Reviewed By: carljm

Differential Revision: D7619628

fbshipit-source-id: 6641cc0bea40c677ae58b38f09b7bb204df52ccb