Eliminate usage of ScheduledSubscriber for requestResponse functionality
Summary:
RSocketRequester's requestResponse method returns a Flowable.
It encapsulates the call of requestResponse method of the RSocketStateMachine with a Flowable and returns it.
Then we immediately subscribe to this result flowable.

In this diff, I get rid of this extra step of creating a flowable and subscribing to it but just call the requestResponse method of the RSocketStateMachine immediately.

Reviewed By: andriigrynenko

Differential Revision: D7752194

fbshipit-source-id: 5fce499803e620f9e80e69bf6d1a5451e2a13ec8