bug fix empty rpc clients that inherit from service
Summary:
I removed an unused constructor for the client_wrapper class in the last diff.
We never used it since we construct the client_wrapper from client.h and not
cython.

Turns out if you define an empty service that inherites from a service.
Then it would produce an cdef cpp_class with no body after the :
So cython would fail to compile. In this case we add "pass" now.

(Note: this ignores all push blocking failures!)

Reviewed By: dddmello

Differential Revision: D7804223

fbshipit-source-id: e856527627f65be2b313850408cd486a1b869ca6