Modify the Thrift Hack compiler to emit "as arraykey" for newtypes used to define legacy enums
Summary: It is common to use enum values (including enums generated from Thrift) as keys into Maps and dicts.  With the changes to the Hack type-checker proposed in D7775754 the latter requires that enums be subtypes of `arraykey`.  This diff modifies the Hack backend of the Thrift compiler to emit `ar arraykey` for newtypes that are used to define legacy enums.

Reviewed By: andrewjkennedy

Differential Revision: D7788769

fbshipit-source-id: db6de67b2fb05c3931b03c1ba613be1dd9506231