Use Empty ResumptionManager to eliminate unnecessary caching
Summary:
When the resumption capability is added to RSocket, its assumed that it will be used extensively. But it's not the case.

If nullptr is provided instead of WarmResumptionManager parameter to the RSocketStateMachine, it internally creates an instance and tracks the frames, instead of assuming that the user doesn't want any resumption.

With this update, we provide an Empty version of WarmResumptionManager to RSocketStateMachine, so it will not create an internal/default version.

The Empty version doesn't cache any frame. So it should bring both memory and cpu usage benefits.

Reviewed By: yfeldblum

Differential Revision: D7842770

fbshipit-source-id: bdb7ff12a20052183f387a398b41e4973f8a5605