Thrift cpp2 - tweak the testing helpers
Summary:
Correct the dependencies of `debug` - it uses `base64_utils` but did not list it as a dependency.

Add `ASSERT_THRIFT_EQ` to match `EXPECT_THRIFT_EQ`.

Add a gmock matcher `ThriftEq`.

Move the ability to rename the left/right objects into the main `debug_output_callback` to avoid having to do that in the matcher, and delete the similar code from `testing.h`.

Use `"\n"` instead of `std::endl` in `debug_output_callback`. This makes it more time-efficient by flushing the stream less often, and fixes a bug that it didn't work unless some other file in the TU already had `#include <ostream>`.

Reviewed By: yfeldblum

Differential Revision: D7868386

fbshipit-source-id: 0178b3e500ced662bf2f42b32c597967707c2802