Update behavior of ref_type fields
Summary:
This change makes the behavior of thrift C++ fields that use `ref="true"`, `cpp.ref_type="..."` or `cpp2.ref_type="..."` more consistent.

Non-optional `ref="true"` and `ref_type="..."` fields will be initialized to point to an object by default. They will point to a cleared (or newly constructed) object when `__clear()` is called on the parent. There is one exception to this - if the type of the field is the same as the enclosing type it will be `nullptr` instead in order to break the infinite recursion which would otherwise occur.

All optional `ref="true"` and `ref_type="..."` fields will be initialized to `nullptr` by default. They will be reset to `nullptr` when `clear()` is called on the parent.

This makes ref_type fields consistent with the treatment of normal fields in that:
* optional fields are now "not set" when default constructed or after a `__clear()`.
* non-optional fields are "set" when default constructed or after a `__clear()`.

Reviewed By: yfeldblum

Differential Revision: D7282591

fbshipit-source-id: bfdb2f66de673b58c348c2215cbbfed48c3451b0