Delete RSRequester and make RSocketClientChannel use RSocketStateMachine
Summary:
RSRequester wraps RSocketStateMachine and forwards the requests to the state machine.
With the recent code updates, RSRequester adds no additional functionality on top of RSocketStateMachine.

This diff gets rid of RSRequester and makes RSocketClientChannel directly use RSocketStateMachine.
Results are the elimination of one more dependency, getting rid of a class, improving the performance.

Reviewed By: alexmalyshev

Differential Revision: D8022397

fbshipit-source-id: 8a0b9a4a0670774c642fd76f04066242af386196