fix crash in Stream cancel with inline executor
Summary:
If the executor the stream is attached is is inline (such as a
SerialExecutor using the default global cpu executor) then this could cause us
to decrement the last reference to sharedState within the rlocked executor
which will cause an assert in SharedMutex because we are destroying it while it
is locked. If we copy the shared_ptr instead of moving it, then we will still
have another reference which outlives the local rExecutor variable.

Reviewed By: phoad

Differential Revision: D8076928

fbshipit-source-id: 6f5127ced85bd6c141cc4614206d3e612ca954df