Thrift: THeader mustn't crash on HTTP response
Summary: Currently when a THeader server receive a valid HTTP response it generates an exception and send it to the client. But since the reception of the HTTP response toggle the clientType to HTTP it try to do it using HTTP instead of THeader, which crash since `httpClientParser_` is uninitialized.

Reviewed By: yfeldblum, phoad

Differential Revision: D7954990

fbshipit-source-id: 0854814b49a7a65c7a49cdbbb253f5aa77e40387