Timeout of one request should not effect others
Summary:
According to the observation, timeout of one [streaming] request does NOT affect the other requests, but itself.

When the streaming request times out on the client side, it just releases the corresponding ThriftClientCallback but it doesn't cancel the corresponding stream on the client side.

Even though the stream is already closed on the server side, it was kept as active on the client side.
So, when the client connection goes away, this stream is thought as active and killed.

With this fix, the stream is canceled as soon as possible when a timeout happens.

Reviewed By: andriigrynenko

Differential Revision: D8247560

fbshipit-source-id: e593549a89fd7a11de5a4dc8cd63ab828385f970