Make RSocketClientChannel hold a shared_ptr to ChannelCounters
Summary: The life-time of the channel and rsocket state machine are not tied to each other (client can e.g. decide to destroy the channel after receiving an error). The state machine currently uses ChannelCounters to signal back to the channel. Making it a shared_ptr is a quick fix to guarantee correctness until we have a cleaner life-time model defined for these classes.

Reviewed By: phoad

Differential Revision: D8306896

fbshipit-source-id: ff761888aff9b000cf9bbf62a80c08eadd3e182f