Gracefully handle service functions that return null streams
Summary:
Service implementors may tend to return empty streams in case of error, instead of returning a stream which provides an error.
```
apache::thrift::Stream<T>  foo() {
   if (errorcase) {
     return {};  // empty Stream
   }
   return realStream;
}
```

In this scenario, it is obvious that they don't care to return a valid error to the client side, but state that there is no data to send down the stream.

This diff changes the behavior of the transport layer, that is not allowing the service function to return empty stream, to allowing it.

Reviewed By: andriigrynenko

Differential Revision: D8307722

fbshipit-source-id: 6d44442a48616ecf284b5aae909113a5874dbe91