Fix edge case in golang thrift codegen for void functions with no exceptions
Summary:
Process() function for requests has been split into Read(), Run() and Write(). Run() returns a result structure, that is passed to Write(), and Write() will call result.Write().

Processor calls Read, Run, and then if we're in a non-oneway function, it calls Write.

The check that was used to decide whether we should return a result should therefore only fail if the function is oneway (i.e. a result is a non-sense), however it also failed if we had no defined exceptions and the function was void. That, in turn, passed <nil> as result to Write(), and write called <nil>.Write(), resulting in a panic.

Fix this by always having run() return a result for non-oneway functions. Since the processor does not call Write() at all for oneway functions, this is safe to do.

Reviewed By: lclarkmichalek

Differential Revision: D8297995

fbshipit-source-id: 3c59c972bf877fab9c40cc63cdae7d8160e36ed8