Unset onDetachable callback of ChannelCounters when the channel is destroyed
Summary:
As the ChannelCounters instance is wrapped with a shared_ptr and shared between StateMachine and the Channel, there is a possibility that ChannelCounters instance can outlive the Channel.

This update unsets onDetachable callback of ChannelCounters when the Channel gets destroyed.
It increases the safety as it prevents accessing to the missing Channel instance through that callback.

Reviewed By: andriigrynenko

Differential Revision: D8319397

fbshipit-source-id: bf341c093e901c387f1064abf83012814c14cb2d