Thrift Serialization: Allow passing in folly::io::QueueAppender
Summary:
The IOBufs used during thrift serialization always use the same growth rate (which is ~16KB) regardless of the eventual size the IOBuf grows too. This can lead to more and fragmented allocations than is necessary for bigger objects.

This change allows the caller of thrift serialization to pass in a folly::io::QueueAppender as the sink of serialized data. The caller can configure the growth rate of the QueueAppender to be close to their expected final serialized size.

Reviewed By: yfeldblum

Differential Revision: D7690820

fbshipit-source-id: 1582097d96b101a86f12dafc761c58adc08169c2