Add the value of the TTransportException.type_ in its message when its value is unknown.
Summary:
Presently when we display the `what()` error message from a `TTransportException` which has an unknown `type_` we only display a default message and hide the actual value of the `type_`.

Since this situation can happen when someone is creating an `TTransportException` by casting an integer into `TTransportExceptionType`, having access to this value can be useful for debugging.

Reviewed By: yfeldblum

Differential Revision: D8264973

fbshipit-source-id: 4045ef3be3a019e0882984198fcb928d7ed95d74