iobuf wrappers to python buffer protocol
Summary:
The idea here is the wrapper class keeps refrences to what ever object owns
the unique_ptr to the folly::iobuf, this maintains that all objects in the chain
are kept alive in python while they are being used, without having to copy the
underlying buffer we use memoryviews for everything.

Also we can accept any object in python that supports the buffer protocol, which
means we could build an iobuf from a numpy array, or build a numpy array from an
iobuf.  We rely on iobuf free fuction to handle sigingling the python buffer
to be released. So it is safe to hand off a python object to be used by C++
outside the gil.

For chains of iobufs we provide a __iter__ which iterates the chains and
produces memoryviews of the underlying buffers.

Reviewed By: ambv

Differential Revision: D8208508

fbshipit-source-id: b9df5d579d8d8635e98033c596982509d4121c08