THeader HTTPClients iobuf chains are out of order
Summary:
IOBuf chain usage within the header generation code appends to the chain rather than prepending, causing the buffer to be injected out of order and it will appear part way through the headers if a very large number of bytes exist in the header. This is seen in any situation where we have > 1 IOBuf in the IOBufQueue. We end up with something like HeaderIOBuf1 -> BodyIOBuf -> HeaderIOBuf2.

This changes the call to prependChain which then ends up with something like HeaderIOBuf1 -> HeaderIOBuf2 -> BodyIOBuf

Reviewed By: yfeldblum

Differential Revision: D8547521

fbshipit-source-id: 5d7c73c058166e410ce2f831958d2ec6e0066885