Moves parser diagnostic message functions into the parsing driver
Summary:
This is the first of a series of diffs that aim at moving all of the stuff in `common.{h|cc}` used by the parser into the parsing driver. For now, the parsing driver lives under `compiler/` since we have not yet isolated the parser-code into `compiler/parse`. But once we do that (with this series of diffs), the parsing driver (along with the flex/bison parsers) would be moved into `compiler/parse`.

This patch, in particular, moves the diagnostic message methods (`pdebug`, `pverbose`, `pwarning`, and `failure`) into the parsing driver. Note that the original methods in `common.cc` shouldn't be removed, as they are still used by the non-parser parts of the compiler (e.g. `main.cc`).

For now I have kept the original behavior of the methods (i.e. printing to `stderr`) to avoid overloading this diff. Eventually they will be changed to more appropriate behaviors (likely to store the diagnostic messages somewhere to be eventually returned to the client of the parser library).

`parse` in `common.{h|cc}` is considered the entry point to the soon-to-be parser library, and will eventually be moved into `compiler/parse`.

Reviewed By: stevegury

Differential Revision: D8594429

fbshipit-source-id: 3c89ce44545daf7b2209f5189641b65adbde6b78