Moves include_file() and friends into the parsing driver
Summary:
(D8594429 for context.)

Moves `include_file()` and its dependencies from `common.{h|cc}` to the parsing driver.

Removed the following globals:

- `g_curdir`
- `g_incl_searchpath`
- `g_program`
- `include_file()`
- `directory_name()`

For RTI team (@[17020033:jmb] and @[1659074342:lego]): Can you help review the `realtime/` part of this diff? This diff eliminates the `g_incl_searchpath` global variable in the thrift compiler. Instead, thrift include search paths are passed in as a parameter to the `parse` function. Currently in `thriftx` and `bifrost`, the search path is set in `main.c`, whereas `parse` is called in `ProgramUtils.h`. The `thriftx` part of this diff is mostly adding plumbing to get the search path options from `main.c` to `ProgramUtils.h`.

Reviewed By: stevegury

Differential Revision: D8598306

fbshipit-source-id: e4f230c36e0c043a4dbf74df10533d46e5b0435a