make union fields returned from getters non-optional
Summary:
The justification is similar to that of D7526831.

Technically, any fields of thrift objects being passed around in python can always be None. However,
the union getter functions actually do assert that the field is non-empty, and it seems reasonable that if the
union's tag is non-empty, we can assume the value is not None. (Otherwise, we have bigger issues to worry about).

Reviewed By: fried

Differential Revision: D8682434

fbshipit-source-id: c39f2a29ef90126230de60b41ae1bda28f0ad747