Moves `parse` entry point into the parsing driver
Summary:
(D8594429 for context.)

(TLDR: Before you call `parse` in `common.{h|cc}`. Now you declare `parsing_driver driver` and do `driver.parse`.)

Up to this point, the entry point into the parsing library has been the `parse()` function in `common.{h|cc}`, which initializes a parsing driver and parses the given program. This diff moves the entry point into the parsing driver itself.

Moreover, it used to be that a new `parsing_driver` is created for each included program. With this change, a single `parsing_driver` would be used for parsing the entire include tree. This allows us to move `global` states that persist across the entire include tree out of `parsing_param`, and instead keep track of them in the `parsing_driver`.

Reviewed By: yfeldblum

Differential Revision: D8789603

fbshipit-source-id: b02f10a65ba3e315e8f65c5526fcac428fd829b3