Improve Enum Performance
Summary:
The python stdlib enums are notoriously slow, mostly .value but some of the
other operations are much slower than normal python objects. We already changed
the code that turns python enums to C++ enums for perf reasons.

This diff replaces the usage of enum.Enum and enum.Flag for all generated code
in thrift-py3.  Since we have access to a Code generator (The thrift Compiler)
We can staticly build cython classes that quack like normal Enum and Flag, and
benchmarks have shown that they will be 6x faster for operations like .value

The static enums replace all thrift enums and the enum backing the Union.Type
The _loadcache on the Union has received a perf improvement, by utilizing the
if/elif to switch statement optimization in cython when using ints, this is
the same improvement we made to python_to_cpp_enum code.

With new Cython based enums, we can utilize cython based type enforcement, so this
required a minnor diff to the compiler to stop excluding enums from hasCythonType?

Improvements to the type defs to support these new enums, and
base classes thrift.py3.types.Enum and thrift.py3.types.Flag are now available
for runtime type determination.

Reviewed By: danshiovitz

Differential Revision: D8775697

fbshipit-source-id: d27064b81a53c056740ff15e012e1ed5fa1669d2