alter debug_equals, add diff-format helper
Summary:
`debug_equals` now invokes the callback for the innermost change only.

The `debug_equals` callback arguments are now pointers and will be `nullptr` when the object or field is missing from that side.

Added a `diff_output_callback` helper and `make_diff_output_callback` to give a fairly minimalist diff-style comparison of two objects.

Reviewed By: yfeldblum

Differential Revision: D8085588

fbshipit-source-id: df50d6e8aecc592e25f89cb6a8546952c9f194cf