Removes the use of `g_curpath`
Summary:
(D8594429 for context.)

`g_curpath` is used for printing out the current thrift file path in the diagnostic messages. However, as said in the previous diff, it is never used in the main part of the compiler, and is always mocked out anyways.

Therefore, rename it to a more appropriate name `g_stage`, and removes the parser library's interaction with it.

Reviewed By: yfeldblum

Differential Revision: D8789639

fbshipit-source-id: c8ceb517c625a4bf607c073fb75cb51a002c7a65