mangle enum members with the name 'name' or 'value', and more duck typing to match enum.Enum
Summary:
Enum values of `name` and `value` are somewhat magical in the stdlib enum, and
one of the reasons they are so slow, we can't easily support them so we add an `_`
to the end of the name if its called `name` or `value`.

Fixed the `dir()` of the enum classes to match std Enum, and added `__members__` to
Union Enums.

Reviewed By: jeffdunn, danshiovitz

Differential Revision: D8831525

fbshipit-source-id: 09a71cb4c70720add7429b3ba73614c244f5f5f5