Per field include_probability in Struct Randomizer
Summary:
Introducing a new struct constraint option: `per_field` that allows struct constrains (e.g. `p_include`) to be scoped at per field level. Currently struct constrains are affecting all fields of a struct. In the following example:
```
struct MyStruct {
  1: i64 a;
  2: i64 b;
}

constraints_for_mystruct = {
  'p_include': 1.0,
}
```
`p_include` refers to the probability (from 0 to 1) of including each non-required field in a struct. `per_field` is now useful if the user wants to set two different probabilities for each field of the struct like this:
```
constraints_for_mystruct = {
  `per_field`: {
    'a': {'p_include': 1.0},
    'b': {'p_include': 0.0}
  }
}
```
This feature is useful in cases that structs can become very complex and we need to control this complexity.

This change is backwards compatible and doesn't change any current behaviour related to the Struct Randomizer.

Reviewed By: paulmar

Differential Revision: D8678663

fbshipit-source-id: dbe36479d184f897046b362cc149f519717661d6