Prevent bad requests from crashing Thrift servers run in duplex mode
Summary:
While attaching a read callback in `Cpp2Channel`, the read callback may fire
immediately, which in turn may lead to the transport being closed. If the
transport is no longer good after attaching the read callback, throw so that
`Cpp2Channel` does not proceed to use the transport assuming that it's still
good.

Reviewed By: yfeldblum, stevegury

Differential Revision: D8879850

fbshipit-source-id: 4574aa716a8be6a0f20d7803f8ca442c9191082d