Fix bug in get_client using unix_path
Summary:
Sometimes we got a unix socket based HeaderChannel that was !good, turns
out we were trying to create a socket to some random garbage in memory.

We were using StringPiece from cython, which is constructed with a char* and a
length, well when that is done Cython just gives a ptr to the buffer of the
bytes object.  Which was not kept alive in the async userdata.

If we pass into the function by C++ string, we get a copy of the bytes buffer.
So then we are safe.

Reviewed By: yfeldblum

Differential Revision: D8939549

fbshipit-source-id: 10f278c7f7b9eda3e9f093d29d3f50d04d76e56a