Moves AST files into a separate library
Summary:
Currently the Thrift AST types are in the `parse` library. It seems like they warrant their own library, since things might depend on the AST but not depend on the parser part.

The types will also be moved into `apache::thrift::compiler` to avoid polluting the global namespace. That will come later in a separate diff.

Reviewed By: yfeldblum

Differential Revision: D8795861

fbshipit-source-id: f7146b5bb32139d21e9e13cb0afd07c5e39cf365