Type usage cleanup
Summary: This is some cleanup on the way some types are referenced to make it easier to do later changes. Specifically, type:flat_name is mostly deprecated and replaced with {{> types/CythonPythonType}}, some values are moved between {{field}} and {{type}} or renamed, etc. The amount of change in the fixtures is minimal - the main fixture change is a custom type for iobufs is no longer used or generated anywhere, since we're directly using __iobuf.IOBuf or equivalent everywhere instead.

Reviewed By: yfeldblum

Differential Revision: D8898024

fbshipit-source-id: a897c83f7b8173e2645c048c98854c9da8077684