Change how Promises are named
Summary: Rename the Promise class - it used to be named Promise_<thrift type> but that doesn't work in the presence of cpp.type'd items, so I switched it to Promise_<cpp name>, since that's the real distinguisher. Also as part of this I'm deprecating type:flat_name except within types/CythonPythonType.mustache - everyone else should use that template and this will ensure custom types are handled correctly.

Reviewed By: yfeldblum

Differential Revision: D9008078

fbshipit-source-id: d7420b67bbb122d15a810356db908be9d3e7ac65