allow function args to be in JSON format
Summary:
We want to eventually pass structures between multiple thrift calls
carried out using `pyremote`.

This requires `pyremote` to accept json as input. And this diff achieves it in
a way that's compatible with previous diff's json output format.

Please note that, this diff also cleans up the current command line options so
that the user can control input format with simply one option. This IS A
NON-BACKWARDS-COMPATIBLE CHANGE to command line interface.

Reviewed By: mingbowan

Differential Revision: D8942481

fbshipit-source-id: 4b238176804b9bf6f0cb05578ee25f7599a23175