generate json template
Summary:
To make it even simpler when using shell to compose a complex series of Thrift
calls, I added another command line option to directly generate a template that
you can modify and pass back into `pyremote` using `-ifmt json`.

It's possible to generate templates for either a Thrift structure, or a function
call.

For function calls, this will generate an entire json containing all the json
forms of the parameters of this function, as well as wrap it correctly in a
top-level object, in the format required by `-ifmt json`.

For Thrift structures, this will only generate the equivalent json form.

Reviewed By: mingbowan

Differential Revision: D8942671

fbshipit-source-id: cdb6ed3862c9d1419f19ada39c08e433c6c6a555