Better support for cpp.type'd binaries
Summary: Add better support for binary types which are cpp.type'd to something else that isn't a folly::IOBuf or unique ptr to one. "Better" in this case means the interface for reading/writing the type is a little more accommodating than std::string. folly::fbstring is excluded from this since it is totally compatible with std::string so there's no need to special-case.

Reviewed By: yfeldblum

Differential Revision: D9031099

fbshipit-source-id: 8022fc556c4f86c12b49189816027950f0fba7f9