Removes output path-related data and logic from t_program
Summary:
Output path is the generator's concern, rather than the AST's. This diff moves them out of `t_program` and into a dedicated `t_generation_context` structure.

Most of the changes are propogating the new `context` parameter down the generator hierarchy, and after that, making the linter happy.

Primary files to review:

- `ast/t_program.{h|cc}`
- `generate/t_generator.{h|cc}`
- `main.cc`
- `py/compiler.cc`
- `py/main.py`
- `py/py_frontend.tcc`

Reviewed By: yfeldblum

Differential Revision: D9141315

fbshipit-source-id: 7152d790aad04a2aff51f6f00c27228e7561a80f