Fixes crash in file-does-not-exist case
Summary:
Currently, if the Thrift parsing library tries to issue a diagnostic message before the parsing begins (e.g. when the file cannot be found), `yyget_text()` will return `NULL`. This `NULL`, given to the `std::string` constructor, then crashes the program.

This diff fixes this bug by using the empty string as the text in this case. Would have used `optional` but the parsing library cannot be on C++17.

Reviewed By: vitaut

Differential Revision: D9366734

fbshipit-source-id: f23c3615b010af46bf6a9141704166e49e006984