Prevent bad requests from crashing Thrift servers run in duplex mode (attempt #2)
Summary:
While attaching a read callback in `Cpp2Channel`, the read callback may fire
immediately, which in turn may lead to the transport being closed. If the
transport is no longer good after attaching the read callback, throw so that
`Cpp2Channel` does not proceed to use the transport assuming that it's still
good.

The first version of this diff had a bug: it is possible that `transport_` is
stolen during the `attachReadCallback()` call, so we need to check it's
non-null before checking `good()`.

Reviewed By: yfeldblum

Differential Revision: D9311064

fbshipit-source-id: ac15399e531e618c102cc7b59d44bb921cfbb5e9