update queued/active requests based on server sampling
Summary: This was a change in behavior in D9067885. observer-queuedRquests only updates the counter if server sampling is enabled, so the counter value itself wasn't changed. The calculation of pendingTaskCount is expensive since it iterates over all the threadpools on threadmanager. This diff reverts back to the original behavior which is calculate the counters only if server sampling is enabled.

Reviewed By: spalamarchuk

Differential Revision: D9417549

fbshipit-source-id: 5e3e486db33db36368b2c4c3edda55f1cf13a0dd